[ review ] Experiment with removing Traceable
Reviewer C suggested that the `Traceable` interface looks to be a
wrapper around a complex `Arbitrary` instance, rather than something
which is required on its own, and that the helper functions are
duplicating existing QC combinators. They are essentially correct:
* `singular` is QC's `pure`
* `choice` is QC's `elements`
* `weighted` is QC's `frequency` in slight disguise.

However, as can be seen in this commit, factoring out `Traceable`
completely leads to an increase in complexity for the type of
`Arbitrary` that then needs implementing (rather than it being provided
so long as `Traceable` is). Specifically, I would argue that requiring
an implicit for the interface definition -- in order to match on it as
part of a function which normally does not take any explicit
arguments -- is unintuitive; that `Traceable` provides a significant
ergonomic benefit. That _is_ different from it being a hard requirement
however, and I will make sure to update the paper to clarify this.

The change would also mean reworking `Arbitrary (Trace <...>)` to not
use the helper function, which should be possible, but I don't think I
have the time for working out the new type variable constraints, show
Idris that the new state chains on into the next `OpRes`, etc. It is
postulated in this commit for now, which allows the file to reload, but
I will be reintroducing both `Traceable` and the `trace` helper
immediately, in order to be able to re-enable the checks that are
commented out in this commit.